requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-localflavor==1.0
gunicorn==19.1.1
httplib2==0.9
mysqlclient==1.3.4
phonenumbers==7.0.1
requests==2.5.0
six==1.8.0
static3==0.5.1
twilio==3.6.13
whitenoise==1.0.5
wsgiref==0.1.2
